+++ junos/resource_static_route_test.go
@@ -83,7 +83,7 @@ func TestAccJunosStaticRoute_basic(t *testing.T) {
 						resource.TestCheckResourceAttr("junos_static_route.testacc_staticRoute_ipv6_instance",
 							"qualified_next_hop.1.next_hop", "2001:db8:85a4::1"),
 						resource.TestCheckResourceAttr("junos_static_route.testacc_staticRoute_ipv6_instance",
-							"qualified_next_hop.1.preference", "st0.0"),
+							"qualified_next_hop.1.interface", "st0.0"),
 						resource.TestCheckResourceAttr("junos_static_route.testacc_staticRoute_ipv6_instance",
 							"community.#", "1"),
 						resource.TestCheckResourceAttr("junos_static_route.testacc_staticRoute_ipv6_instance",
@@ -158,8 +158,8 @@ resource junos_static_route testacc_staticRoute_instance {
     metric     = 101
   }
   qualified_next_hop {
-	next_hop   = "192.0.2.250"
-	interface  = "st0.0"
+    next_hop   = "192.0.2.250"
+    interface  = "st0.0"
   }
   community = ["no-advertise"]
 }
@@ -199,9 +199,9 @@ resource junos_static_route testacc_staticRoute_ipv6_instance {
     metric = 101
   }
   qualified_next_hop {
-  next_hop   = "2001:db8:85a4::1"
-  interface  = "st0.0"
-}
+    next_hop   = "2001:db8:85a4::1"
+    interface  = "st0.0"
+  }
   community = ["no-advertise"]
 }
 `